requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
mysql-connector-python  # Or psycopg2-binary for PostgreSQL, or remove if using SQLite
pydantic[email]
pydantic-settings # Added for robust settings management
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv
email_validator
python-multipart
Authlib
requests # Added for making HTTP requests, e.g., for Google OAuth profile image download
pydantic_settings
aiofiles